Image Processing Full Stack/KidneyDiseaseDetector/requirements.txt
flask>=3.0.0
ultralytics
opencv-python
numpy
Pillow
torch
torchvision
werkzeug
